[BUGFIX] Reflect SQL "limit" when showing translation overlays
Both, the core backend list view as well as the frontend rendering
use a "limit" when it comes to fetching translation overlays. This
resuts in only showing the very first translation for a certain
contant element.
The implementation of nxcondensedbelayout needs to reflect this
but used to use the latest content instead.