TASK: Don't overrule tt_content::sys_language_uid when copying records
If an existing record is pasted into a specific position, Gridelements
overrules the original records language by the one currently active in
the backend page view.

Although that might be an important feature when moving records around
amongst containers of different languages that is the rare case in the
real world. Pretty often containers are not of type "language > 0" but
of type "language = -1", meaning it's not at ranslated container.

Updating such a container from "langauge = -1" to "language = 0" while
copying just doesn't make much sense.