BUGFIX: Now "paste record into container" makes all crated records target the same slot
When using the "paste record" mechanism of the BE ui, that's actually
two datahandler operations in a row. The first creates pure copy records,
the second adjusts colpos.
This change makes the second operation apply not only to the copied
record but to all automatically created translation records as well.

Fixed behavior: When a tt_content was copied without this change and
its copy source record got deleted afterwards, the source container
appeared to be empty in BE although tt_contents pointed to it. Deleting
this seemingly empty container traversed through child records and
removed content that was not supposed to be removed.