TASK: Detect translation by l10n_source, not l18n_parent
The l18n_parent means "the record I am the translation of", where
the l10n_surce means "I contain corresponding data in another language,
but I'm not a translation. The difference is that records having both,
l10n_source and l18n_parent point to the same record are meant to replace
content but those records having l10n_source point but l18n_parent not
point to a source record are meant to move freely and potentially appear
side by side.

This acutally means: Using l18n_parent is the intended behavior. But
since having l10n_source and l18n_parent point to different content records
is contradicting the intention of this very extension we need to somehow
avoid this.

Always relying on l10n_source does this job. If l18n_parent and l10n_source
are equal, this change acutally doesn't change a thing. If l18n_parent is
not set but l10n_source is, this change mixed-mode content in both, fe
and be.