TASK: Keep translations soting in sync with default record
Those columns should be kept in sync by DataHandler when tt_content
is updated in the backend.
But on one hand, sometimes database updates don't respect the TCA
setting.
And on the other hand, "sorting" almost never is in sync, which is
the primary use case here.

Keeping sorting in sync in FE is required since that's what the BE
mode of condensedlayout does anyway.