BUGFIX: Use l18n_parent as language pointer, not l10n_source
The field value of l18n_parent means "the record I am the translation
of", where the field value of l10n_source means "I contain corresponding
data in another language, but I'm not a translation."

The purpose of this extension is to basically have both fields in sync.
The TYPO3 so called "mixed mode", where l10n_source and l18n_parent point
to different sources, is usually neither intended nor anderstood by
editors.

In a previous commit I made use of the l10n_source field for backend
behavior because that seemd as a better fit.

This is not true because frontend behaves ot based on l10n_source but
based on l18n_parent. In fact, the l10n_source is *only* used for
mixed mode backend behavior.

This patch now switches the condensed vackend layout back to using
l18n_parent and provides an install tool update step to set those in
sync.