BUGFIX: Keep grid container and column of translations in sync
When moving translated content into containers or out of containers,
the grid elements column and container values of the translations
were not updated. So technically the translation didn't get moved
into that container but remained outside.

This causes problems when a page is copied: The translation passes
the data handler twice, once for it being a translation of an inner
element and a second time for it being not inside of a container. This
results in clumsy behavior when it comes to IRRE like e.g. files.

This patch now keeps those columns in sync when a record is moved.